+++ components/operations/utils-operations/check-address-data.tsx
@@ -13,37 +13,14 @@ import { OperationsSubmitButton } from '@/components/operations/operations-submi
 import { useState } from 'react';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { AlertCircle } from 'lucide-react';
-import { OperationsRadioGroup } from '@/components/operations/operations-radio-group';
 import axios, { AxiosError } from 'axios';
 import { useConfig } from '@useelven/core';
 
-const tokenTypes: { name: string; description: string }[] = [
-  {
-    name: 'fungible',
-    description: 'Fungible ESDT',
-  },
-  {
-    name: 'non-fungible',
-    description: 'Non-fungible (NFT)',
-  },
-  {
-    name: 'semi-fungible',
-    description: 'Semi-fungible (SFT)',
-  },
-  {
-    name: 'meta',
-    description: 'Meta ESDT',
-  },
-];
-
 const formSchema = z.object({
-  tokenId: z.string().min(1, 'The field is required'),
-  tokenType: z.enum(['fungible', 'non-fungible', 'semi-fungible', 'meta'], {
-    required_error: 'Please choose the type of the token',
-  }),
+  addressOrHerotag: z.string().min(1, 'The field is required'),
 });
 
-export const CheckTokenData = () => {
+export const CheckAddressData = () => {
   const [result, setResult] = useState<string>();
   const [pending, setPending] = useState(false);
   const { apiAddress } = useConfig();
@@ -51,28 +28,20 @@ export const CheckTokenData = () => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      tokenId: '',
-      tokenType: 'fungible',
+      addressOrHerotag: '',
     },
   });
 
-  const onSubmit = async ({
-    tokenId,
-    tokenType,
-  }: z.infer<typeof formSchema>) => {
-    let endpointFragment;
-
-    if (tokenType === 'fungible') {
-      endpointFragment = 'tokens';
-    } else {
-      endpointFragment = 'collections';
-    }
+  const onSubmit = async ({ addressOrHerotag }: z.infer<typeof formSchema>) => {
+    const endpointFragment = !addressOrHerotag.startsWith('erd')
+      ? 'usernames'
+      : 'accounts';
 
     try {
       // TODO: replace with useElven useApiCall when ready to handle such cases
       setPending(true);
       const result = await axios.get(
-        `${apiAddress}/${endpointFragment}/${tokenId.trim()}`,
+        `${apiAddress}/${endpointFragment}/${addressOrHerotag.trim()}`,
         {
           headers: {
             'Content-Type': 'application/json',
@@ -92,32 +61,26 @@ export const CheckTokenData = () => {
   return (
     <>
       <DialogHeader className="p-8 pb-0">
-        <DialogTitle>Check token data</DialogTitle>
+        <DialogTitle>Check address data</DialogTitle>
         <DialogDescription>
-          Check basic token data. It can be a standard fungible token or
-          collection id for NFT/SFT/Meta tokens. Useful because Explorer API has
-          agressive cache in some cases.
+          Check basic address data (also using a herotag). It can also be a
+          smart contract address. Useful because Explorer API has agressive
+          cache in some cases.
         </DialogDescription>
       </DialogHeader>
       <div className="overflow-y-auto py-0 px-8">
         <Form {...form}>
           <form
-            id="check-token-data-form"
+            id="check-address-data-form"
             onSubmit={form.handleSubmit(onSubmit)}
             className="space-y-8"
           >
             <div className="flex-1 overflow-auto p-1">
-              <OperationsRadioGroup
-                items={tokenTypes}
-                name="tokenType"
-                label="Token type"
-                description="Please choose the type of the token to check."
-              />
               <OperationsInputField
-                name="tokenId"
-                label="Token id"
-                placeholder="Example: MyToken-23432"
-                description="Please provide your token id. Fungible or collection id."
+                name="addressOrHerotag"
+                label="Wallet address or herotag"
+                placeholder="Example: erd1.... or <herotag_name>"
+                description="Please provide a wallet address to check. You can also use a herotag or you can check a smart contract address."
               />
             </div>
           </form>
@@ -125,7 +88,7 @@ export const CheckTokenData = () => {
         {result && (
           <Alert className="w-full overflow-y-auto">
             <AlertCircle className="h-4 w-4" />
-            <AlertTitle>The token data:</AlertTitle>
+            <AlertTitle>The address data:</AlertTitle>
             <AlertDescription className="mt-3 w-full">
               <pre>
                 <code>{result}</code>
@@ -136,7 +99,7 @@ export const CheckTokenData = () => {
       </div>
       <DialogFooter className="py-4 px-8">
         <OperationsSubmitButton
-          formId="check-token-data-form"
+          formId="check-address-data-form"
           pending={pending}
           isPublic
         />