+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [0.3.0](https://github.com/xdevguild/buildo.dev/releases/tag/v0.3.0) (2023-10-23)
+ - add herotag functionality
+ - add check address (also by herotag) utility
+
 ### [0.2.0](https://github.com/xdevguild/buildo.dev/releases/tag/v0.2.0) (2023-10-22)
 - add transfer creation role for NFT/SFT/Meta
 - add stop creation for NFT/SFT/Meta